204: Move `if the collision is of size 2 and one of the keys is removed, the remaining key must be promoted` test
