+++ core/commonTest/src/stress/map/PersistentHashMapTest.kt
@@ -6,7 +6,6 @@
 package tests.stress.map
 
 import kotlinx.collections.immutable.PersistentMap
-import kotlinx.collections.immutable.implementations.immutableMap.PersistentHashMap
 import kotlinx.collections.immutable.persistentHashMapOf
 import tests.NForAlgorithmComplexity
 import tests.distinctStringValues
@@ -350,26 +349,6 @@ class PersistentHashMapTest : ExecutionTimeMeasuringTest() {
                 }
             }
         }
-    }
-
-    @Test
-    fun `if the collision is of size 2 and one of the keys is removed, the remaining key must be promoted`() {
-        val map1: PersistentHashMap<Int, String> =
-            persistentHashMapOf(-1 to "a", 0 to "b", 32 to "c") as PersistentHashMap<Int, String>
-        val builder = map1.builder()
-        val map2 = builder.build()
-
-        assertTrue(map1.equals(builder))
-        assertEquals(map1, map2.toMap())
-        assertEquals(map1, map2)
-
-        val map3 = map1.remove(0)
-        builder.remove(0)
-        val map4 = builder.build()
-
-        assertTrue(map3.equals(builder))
-        assertEquals(map3, map4.toMap())
-        assertEquals(map3, map4)
     }
 
     private fun testAfterOperation(
